supply-chain-dashboard/requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
statsmodels
scikit-learn
openpyxl
google-generativeai
